requirements.txt
customtkinter
faster-whisper
SpeechRecognition
ffmpeg-python
deep-translator
darkdetect
pillow
